Update triggerReferralSteps logic
